Fix a compilation error

+++ taglib/toolkit/tfile.h
@@ -293,7 +293,7 @@ namespace TagLib {
     /*!
      * Returns the buffer size that is used for internal buffering.
      */
-    static uint bufferSize();
+    static size_t bufferSize();
 
   private:
     File(const File &);
